+++ tekore/_client/api/track.py
@@ -1,5 +1,5 @@
 from ..base import SpotifyBase
-from ..decor import send_and_process
+from ..decor import send_and_process, scopes
 from ..process import single, model_list
 from ..chunked import chunked, join_lists
 from tekore.model import FullTrack, AudioFeatures, AudioAnalysis, ModelList
@@ -8,6 +8,7 @@ from tekore.model import FullTrack, AudioFeatures, AudioAnalysis, ModelList
 class SpotifyTrack(SpotifyBase):
     """Track API endpoints."""
 
+    @scopes()
     @send_and_process(single(FullTrack))
     def track(
             self,
@@ -31,6 +32,7 @@ class SpotifyTrack(SpotifyBase):
         """
         return self._get('tracks/' + track_id, market=market)
 
+    @scopes()
     @chunked('track_ids', 1, 50, join_lists)
     @send_and_process(model_list(FullTrack, 'tracks'))
     def tracks(
@@ -55,6 +57,7 @@ class SpotifyTrack(SpotifyBase):
         """
         return self._get('tracks/?ids=' + ','.join(track_ids), market=market)
 
+    @scopes()
     @send_and_process(single(AudioAnalysis))
     def track_audio_analysis(self, track_id: str) -> AudioAnalysis:
         """
@@ -70,6 +73,7 @@ class SpotifyTrack(SpotifyBase):
         """
         return self._get('audio-analysis/' + track_id)
 
+    @scopes()
     @send_and_process(single(AudioFeatures))
     def track_audio_features(self, track_id: str) -> AudioFeatures:
         """
@@ -82,6 +86,7 @@ class SpotifyTrack(SpotifyBase):
         """
         return self._get('audio-features/' + track_id)
 
+    @scopes()
     @chunked('track_ids', 1, 100, join_lists)
     @send_and_process(model_list(AudioFeatures, 'audio_features'))
     def tracks_audio_features(self, track_ids: list) -> ModelList: