+++ tekore/_client/api/user.py
@@ -1,12 +1,14 @@
 from ..base import SpotifyBase
-from ..decor import send_and_process
+from ..decor import send_and_process, scopes
 from ..process import single
+from tekore._auth import scope
 from tekore.model import PublicUser, PrivateUser
 
 
 class SpotifyUser(SpotifyBase):
     """User API endpoints."""
 
+    @scopes()
     @send_and_process(single(PublicUser))
     def user(self, user_id: str) -> PublicUser:
         """
@@ -24,14 +26,14 @@ class SpotifyUser(SpotifyBase):
         """
         return self._get('users/' + user_id)
 
+    @scopes(optional=[scope.user_read_private, scope.user_read_email])
     @send_and_process(single(PrivateUser))
     def current_user(self) -> PrivateUser:
         """
         Get current user's profile.
 
-        Requires the user-read-private scope to return
-        user's country and product subscription level.
-        Requires the user-read-email scope to return user's email.
+        The user-read-private scope allows the user's country and
+        product subscription level to be returned.
 
         Returns
         -------